Some refactoring of admin controllers
This is a bit of a mixed bag. Mainly it's just moving the admin action/controllers into the library autoload directory.
However at the same time i've fixed a couple of minor bugs, to do with dashboard widgets and updated some stuff to allow for future refactoring regarding request stuff.
Don't expect psr compliant stuff just yet !
