+++ admin/includes/auto_loaders/ajax.core.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @package admin
- * @copyright Copyright 2003-2014 Zen Cart Development Team
+ * @copyright Copyright 2003-2015 Zen Cart Development Team
  * @license http://www.zen-cart.com/license/2_0.txt GNU Public License V2.0
  * @version $Id:
  */
@@ -72,9 +72,9 @@ if (!defined('USE_PCONNECT')) define('USE_PCONNECT', 'false');
   $autoLoadConfig[0][] = array('autoType'=>'classInstantiate',
                                'className'=>'zcPassword',
                                'objectName'=>'zcPassword');
-  $autoLoadConfig[0][] = array('autoType'=>'class',
-                               'loadFile'=>'Request.php',
-                               'classPath'=>DIR_CATALOG_LIBRARY . 'zencart/platform/src/');
+  $autoLoadConfig[0][] = array('autoType'=>'classInstantiate',
+                             'className'=>'\\ZenCart\\Platform\\Request',
+                             'objectName'=>'zcRequest');
 
 /**
  * Breakpoint 10.
