+++ includes/autoload_namespaces.php
@@ -17,6 +17,8 @@ if (!defined('DIR_CATALOG_LIBRARY')) {
  * An array of namespace => basedir configurations
  */
 return array(
+    '\ZenCart\Admin\Controllers' => DIR_CATALOG_LIBRARY . 'zencart/admin/Controllers/src',
+    '\ZenCart\Admin\AjaxDispatch' => DIR_CATALOG_LIBRARY . 'zencart/admin/AjaxDispatch/src',
     '\ZenCart\Admin\DashboardWidget' => DIR_CATALOG_LIBRARY . 'zencart/admin/DashboardWidget/src',
     '\Aura\Web' => DIR_CATALOG_LIBRARY . 'aura/web/src',
     'ZenCart\Platform' => DIR_CATALOG_LIBRARY. 'zencart/platform/src/',
